requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-huggingface>=0.0.1
llama-index>=0.9.0
pymilvus>=2.3.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2

# File processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pillow>=10.0.0
pytesseract>=0.3.10
pydub>=0.25.1
moviepy>=1.0.3
beautifulsoup4>=4.12.2
requests>=2.31.0

# Web scraping
scrapy>=2.11.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.1
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2

# Testing
pytest>=7.4.3